Enable multithreading capabilities for LAMMPS
Added the ability for sim.run() to read in the number of threads (the number of cores to be used) and the mpi prefix (e.g., mpiexec). Running serial is still available if desired with default settings.
In this fix np and prefix are defined in mc_md which is non-standard in the current stable pysimm release. A future fix should enable np and prefix to be added as dictionary entries to md_props to fit better into pysimm documentation.